requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
idna==3.4
mysql-connector==2.2.9
mysqlclient==2.1.1
Pillow==9.2.0
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.12.2
tzdata==2022.2
urllib3==1.26.13
